(refactor) Move component tests to its own folder
